Part 4 initial created project, removed the old one that was pushed on my branch
