+++ G3_MicroProject/Part4/app/build.gradle
@@ -4,7 +4,7 @@ android {
     compileSdkVersion 29
     buildToolsVersion "29.0.3"
     defaultConfig {
-        applicationId "com.example.part3"
+        applicationId "com.example.g3.graphicalguessinggame"
         minSdkVersion 29
         targetSdkVersion 29
         versionCode 1
@@ -26,5 +26,4 @@ dependencies {
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test.ext:junit:1.1.0'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.1'
-    implementation 'androidx.gridlayout:gridlayout:1.0.0'
 }
