+++ G3_MicroProject/Part4/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.part4">
+    package="com.example.g3.graphicalguessinggame">
 
     <uses-permission android:name="android.permission.VIBRATE" />
     <application
@@ -10,8 +10,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name="com.example.part4.GuessingGame"></activity>
-        <activity android:name="com.example.part4.MainActivity">
+        <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
