--- G3_MicroProject/Part4/app/src/main/java/com/example/part4/desktop.ini
@@ -1,2 +1,0 @@
-[.ShellClassInfo]
-LocalizedResourceName=part4
