--- G3_MicroProject/Part4/app/src/main/res/drawable/song_selector.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item
-        android:state_selected="true"
-        android:drawable="@drawable/textBG">
-    </item>
-
-</selector>
