+++ G3_MicroProject/Part4/app/src/main/res/values/colors.xml
@@ -3,7 +3,4 @@
     <color name="colorPrimary">#008577</color>
     <color name="colorPrimaryDark">#00574B</color>
     <color name="colorAccent">#D81B60</color>
-
-    <drawable name="textBG">#0125FF</drawable>
-
 </resources>
