+++ G3_MicroProject/Part4/app/src/main/res/values/strings.xml
@@ -1,12 +1,3 @@
 <resources>
-    <string name="app_name">Progress Bar</string>
-    <string name="Score">Your score is: </string>
-    <string name="Chance">Remaining chances are:\n</string>
-    <string name="winMessage">CONGRATULATION! You\'ve won</string>
-    <string name="loseMessage">SORRY GAME OVER</string>
-    <string name="greatMessage">The number you have entered is greater than the number chosen by the program</string>
-    <string name="lessMessage">The number you have entered is less than the number chosen by the program</string>
-    <string name="timeover">Sorry, game over, your time is exhausted</string>
-    <string name="guessMessage">Guess the number generated by the app between \n0 to 100</string>
-    <string name="invalidNumber">Enter a valid number!</string>
+    <string name="app_name">Graphical Guessing Game</string>
 </resources>
