+++ G3_MicroProject/Part4/settings.gradle
@@ -1,2 +1,2 @@
 include ':app'
-rootProject.name='Part3'
+rootProject.name='Graphical Guessing Game'
